@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,7 +27,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  google_fonts: ^2.1.0
+  google_fonts: ^2.3.1
   flutter_screenutil: ^5.0.1
 
 dev_dependencies:
@@ -48,9 +48,13 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-
+  #   - images/logo.png
+  fonts:
+    - family: mak
+      fonts:
+        - asset: fonts/mak.ttf
+          weight: 300
+          
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
 
@@ -63,10 +67,7 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
 
-  fonts:
-    - family: Mak       
-      fonts:
-        - asset: fonts/mak.ttf
+  
         
   #
   # For details regarding fonts from package dependencies,
