@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:sovmestno/modules/form/form_widgets/address_field_widget.dart';
-import 'package:sovmestno/modules/form/form_widgets/checkbox_widget.dart';
-import 'package:sovmestno/widgets/custom_button_widget.dart';
+import 'package:sovmestno/presentation/registration/forms/form_widgets/address_field_widget.dart';
+import 'package:sovmestno/presentation/registration/forms/form_widgets/checkbox_widget.dart';
+import 'package:sovmestno/widgets/buttons/custom_main_button.dart';
+
 
 class MainForm extends StatelessWidget {
   final VoidCallback onSavePressed;
-  const MainForm({Key key,@required this.onSavePressed }) : super(key: key);
+  const MainForm({Key? key,required this.onSavePressed }) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -23,8 +24,8 @@ class MainForm extends StatelessWidget {
         const AddressField(),
         const SizedBox(height: 46),
         CustomButtonWidget(
-            text: 'Сохранить',
-            onPressed: () {
+            title: 'Сохранить',
+            callback: () {
               debugPrint('Сохранить');
               onSavePressed.call();
             }),
