@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:sovmestno/constants/colors.dart';
-import 'package:sovmestno/modules/registration/registration_widgets/log_into_account_widget.dart';
-import 'package:sovmestno/modules/registration/registration_widgets/register_button_widget.dart';
 
 
 class UserImageWidget extends StatelessWidget {
-  const UserImageWidget({Key key}) : super(key: key);
+  const UserImageWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -30,7 +28,7 @@ class UserImageWidget extends StatelessWidget {
               child: ElevatedButton(
                 child: const Text(
                   'Загрузить фотографию',
-                  style: TextStyle(color: MyColors.backgroundButton),
+                  style: TextStyle(color: AppColors.backgroundButton),
                 ),
                 onPressed: () {},
                 style: ElevatedButton.styleFrom(
@@ -40,7 +38,7 @@ class UserImageWidget extends StatelessWidget {
                     borderRadius: BorderRadius.circular(6.0),
                     side: const BorderSide(
                       width: 1.0,
-                      color: MyColors.backgroundButton,
+                      color: AppColors.backgroundButton,
                     ),
                   ),
                 ),
@@ -51,12 +49,9 @@ class UserImageWidget extends StatelessWidget {
         const SizedBox(height: 18),
         const Text(
           'Максимальный размер фото  1 MB.',
-          style: TextStyle(fontSize: 12, color: MyColors.backgroundButton),
+          style: TextStyle(fontSize: 12, color: AppColors.backgroundButton),
         ),
         const SizedBox(height: 37),
-        const RegisterButtonWidget(),
-        const SizedBox(height: 16),
-        const LogIntoAccountWidget()
       ],
     );
   }
