@@ -1,26 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:sovmestno/constants/colors.dart';
+import 'package:sovmestno/modules/registration/registration_widgets/login_by_email_widget.dart';
 import 'package:sovmestno/modules/registration/registration_widgets/registration_app_bar.dart';
-import 'package:sovmestno/modules/registration/registration_widgets/registration_widget.dart';
 import 'package:sovmestno/widgets/back_button_widget.dart';
 
-class LoginByEmail extends StatefulWidget {
-  const LoginByEmail({Key key}) : super(key: key);
+class LoginByEmailScreen extends StatefulWidget {
+  const LoginByEmailScreen({Key key}) : super(key: key);
 
   @override
-  _LoginByEmailState createState() => _LoginByEmailState();
+  _LoginByEmailScreenState createState() => _LoginByEmailScreenState();
 }
 
-class _LoginByEmailState extends State<LoginByEmail> {
+class _LoginByEmailScreenState extends State<LoginByEmailScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: MyColors.appBarColor,
-      appBar: const RegistrationAppBar(title: 'Совместно'),
+      appBar: const RegistrationAppBar(),
       body: SingleChildScrollView(
         scrollDirection: Axis.vertical,
         child: Padding(
-          padding: const EdgeInsets.only(left: 104, right: 104),
+          padding: EdgeInsets.symmetric(horizontal: 104.w),
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -39,11 +40,11 @@ class _LoginByEmailState extends State<LoginByEmail> {
               const SizedBox(height: 19),
               BackButtonWidget(
                 onPressed: () {
-                  Navigator.pushNamed(context, '/authorization');
+                  Navigator.pushNamed(context, '/FormMainScreen');
                 },
               ),
               const SizedBox(height: 44),
-              const Registration(),
+              const LoginByEmailWidget(),
               const SizedBox(height: 32),
             ],
           ),
